LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-inherited-expected.txt


FAIL .child 1 assert_equals:
<div class="child" data-offset-x="20" data-offset-y="10" data-expected-width="100" data-expected-height="100"></div>
width expected 100 but got 0
